requirements.txt
google-generativeai
Pillow
pandas
